+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.21
+### Added
+
+## fixed
+- Fixed handling of Oauth (Bearer token).
+
 ## 0.3.20
 ### Added
 - Length for Column "SUB_PROTOCOL_BODY" has been extended to 2048.